move Compactor to own file
