+++ libsqlx-server/src/allocation/primary/mod.rs
@@ -16,6 +16,8 @@ use crate::meta::DatabaseId;
 
 use super::{ConnectionHandler, ExecFn, FRAMES_MESSAGE_MAX_COUNT};
 
+pub mod compactor;
+
 const MAX_STEP_BATCH_SIZE: usize = 100_000_000; // ~100kb
                                                 //
 pub struct PrimaryDatabase {
